add preconnect
