+++ blocks/embed/lite-vimeo-embed/lite-vimeo-embed.js
@@ -1,4 +1,5 @@
 /* eslint-disable  */
+import { addPrefetch } from "../../../libs/utils/utils";
 
 /**
  * adapted from https://github.com/luwes/lite-vimeo-embed
@@ -112,20 +113,6 @@ class LiteVimeo extends (globalThis.HTMLElement ?? class {}) {
 
 if (globalThis.customElements && !globalThis.customElements.get('lite-vimeo')) {
   globalThis.customElements.define('lite-vimeo', LiteVimeo);
-}
-
-/**
- * Add a <link rel={preload | preconnect} ...> to the head
- */
-function addPrefetch(kind, url, as) {
-  const linkElem = document.createElement('link');
-  linkElem.rel = kind;
-  linkElem.href = url;
-  if (as) {
-    linkElem.as = as;
-  }
-  linkElem.crossorigin = true;
-  document.head.append(linkElem);
 }
 
 /**
